0429_assignment/pyproject.toml
[project]
name = "storybook-agent"
version = "0.1.0"
description = "Two-agent children's storybook generator (Story Writer + Illustrator) built on Google ADK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.12.0",
    "google-genai>=1.31.0",
    "litellm>=1.76.0",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
]
